+++ example/sample-projects/with-configure-step/run.t
@@ -2,8 +2,8 @@
          ocaml config.full
       ocamldep src/.plop.eobjs/config.ml.d
       ocamldep src/.plop.eobjs/plop.ml.d
-        ocamlc src/.plop.eobjs/config.{cmi,cmo,cmt}
-        ocamlc src/.plop.eobjs/plop.{cmi,cmo,cmt}
-      ocamlopt src/.plop.eobjs/plop.{cmx,o}
-      ocamlopt src/.plop.eobjs/config.{cmx,o}
+        ocamlc src/.plop.eobjs/byte/config.{cmi,cmo,cmt}
+        ocamlc src/.plop.eobjs/byte/plop.{cmi,cmo,cmt}
+      ocamlopt src/.plop.eobjs/native/plop.{cmx,o}
+      ocamlopt src/.plop.eobjs/native/config.{cmx,o}
       ocamlopt src/plop.exe