+++ src/inline_tests.ml
@@ -168,9 +168,10 @@ include Sub_system.Register_end_point(
         sprintf "%s.inline-tests" (Lib_name.Local.to_string (snd lib.name))
       in
 
-      let inline_test_dir = Path.relative dir ("."^inline_test_name) in
+      let inline_test_dir = Path.relative dir ("." ^ inline_test_name) in
 
-      let obj_dir = Obj_dir.make_exe ~dir:inline_test_dir inline_test_name in
+      let obj_dir =
+        Obj_dir.make_exe ~dir:inline_test_dir ~name:inline_test_name in
 
       let name = "run" in
       let main_module_filename = name ^ ".ml" in