+++ src/utils.mli
@@ -22,7 +22,13 @@ val library_object_directory
   -> Lib_name.Local.t
   -> Path.t
 
-val library_private_obj_dir : obj_dir:Path.t -> Path.t
+(** cmx, .a *)
+val library_native_dir     : obj_dir:Path.t -> Path.t
+
+(** cmo, cmi, cmt, cmti *)
+val library_byte_dir       : obj_dir:Path.t -> Path.t
+val library_public_cmi_dir : obj_dir:Path.t -> Path.t
+val library_private_dir    : obj_dir:Path.t -> Path.t
 
 (** Return the directory where the object files for the given
     executable should be stored. *)