+++ test/blackbox-tests/test-cases/dune-ppx-driver-system/run.t
@@ -80,7 +80,7 @@ Test that going throught the -ppx option of the compiler works
       ocamldep .test_ppx_staged.objs/test_ppx_staged.ml.d
   tool name: ocamldep
   args:--as-ppx --cookie library-name="test_ppx_staged"
-        ocamlc .test_ppx_staged.objs/test_ppx_staged.{cmi,cmo,cmt}
+        ocamlc .test_ppx_staged.objs/byte/test_ppx_staged.{cmi,cmo,cmt}
   tool name: ocamlc
   args:--as-ppx --cookie library-name="test_ppx_staged"
 
@@ -160,17 +160,17 @@ Test using installed drivers
 
   $ dune build --display short --root jbuild-driver @all
   Entering directory 'jbuild-driver'
-        ocamlc .testdriver.objs/testdriver.{cmi,cmo,cmt}
+        ocamlc .testdriver.objs/byte/testdriver.{cmi,cmo,cmt}
         ocamlc testdriver.cma
-      ocamlopt .testdriver.objs/testdriver.{cmx,o}
+      ocamlopt .testdriver.objs/native/testdriver.{cmx,o}
       ocamlopt testdriver.{a,cmxa}
       ocamlopt testdriver.cmxs
       ocamlopt .ppx/jbuild/631757a4a4789e0bd29628f7a73480f7/ppx.exe
            ppx test_ppx_args.pp.ml
       ocamldep .test_ppx_args.objs/test_ppx_args.pp.ml.d
-        ocamlc .test_ppx_args.objs/test_ppx_args.{cmi,cmo,cmt}
+        ocamlc .test_ppx_args.objs/byte/test_ppx_args.{cmi,cmo,cmt}
         ocamlc test_ppx_args.cma
-      ocamlopt .test_ppx_args.objs/test_ppx_args.{cmx,o}
+      ocamlopt .test_ppx_args.objs/native/test_ppx_args.{cmx,o}
       ocamlopt test_ppx_args.{a,cmxa}
       ocamlopt test_ppx_args.cmxs