+++ test/blackbox-tests/test-cases/formatting/run.t
@@ -5,14 +5,14 @@ Formatting can be checked using the @fmt target:
   $ dune build --display short @fmt
       ocamldep fake-tools/.ocamlformat.eobjs/ocamlformat.ml.d
       ocamldep fake-tools/.ocamlformat.eobjs/refmt.ml.d
-        ocamlc fake-tools/.ocamlformat.eobjs/ocamlformat.{cmi,cmo,cmt}
-      ocamlopt fake-tools/.ocamlformat.eobjs/ocamlformat.{cmx,o}
+        ocamlc fake-tools/.ocamlformat.eobjs/byte/ocamlformat.{cmi,cmo,cmt}
+      ocamlopt fake-tools/.ocamlformat.eobjs/native/ocamlformat.{cmx,o}
       ocamlopt fake-tools/ocamlformat.exe
    ocamlformat enabled/.formatted/ocaml_file.ml
   File "enabled/ocaml_file.ml", line 1, characters 0-0:
   Files _build/default/enabled/ocaml_file.ml and _build/default/enabled/.formatted/ocaml_file.ml differ.
-        ocamlc fake-tools/.ocamlformat.eobjs/refmt.{cmi,cmo,cmt}
-      ocamlopt fake-tools/.ocamlformat.eobjs/refmt.{cmx,o}
+        ocamlc fake-tools/.ocamlformat.eobjs/byte/refmt.{cmi,cmo,cmt}
+      ocamlopt fake-tools/.ocamlformat.eobjs/native/refmt.{cmx,o}
       ocamlopt fake-tools/refmt.exe
          refmt enabled/.formatted/reason_file.rei
   File "enabled/reason_file.rei", line 1, characters 0-0: