+++ test/blackbox-tests/test-cases/ocamldep-multi-stanzas/run.t
@@ -26,8 +26,8 @@
   to a single library or executable.
   This warning will become an error in the future.
       ocamldep src/.a.objs/x.ml.d
-        ocamlc src/.a.objs/a.{cmi,cmo,cmt}
-        ocamlc src/.a.objs/a__X.{cmi,cmo,cmt}
+        ocamlc src/.a.objs/byte/a.{cmi,cmo,cmt}
+        ocamlc src/.a.objs/byte/a__X.{cmi,cmo,cmt}
         ocamlc src/a.cma
 
   $ dune exec ./test.exe --debug-dep --display short --root dune